requirements.txt
flask
playwright
gspread
google-auth